requirements.txt
requests
python-dotenv
schedule
websocket-client